adjust date filter in blog posts
makes sure that the default date when entering /blog is the last day of the month at 23:59:59 taking into account time zone differences. Also when selecting a new date with month picker makes sure that the selected date is the last day of the selected month at 23:59:59 also checking with Argentinian time zones.